refactor topic page
